refactor(schema): del infra datacenter/room props
Thee datacenter and room properties of Infrastructure object were
actually wrong as InfratructurePart is actually attached to a rack which
could be in any datacenter/room.

The correct way to find out the rooms and datacenters of an
infrastructure is look over the rack in the parts of its layout.