docs: initial version
This commit adds documentation in Antora and asciidoctor format for
most aspects of RacksDB except the quickstart guide, as no release is
available to install yet.

There is a Python script to generate the database structure reference
documentation based on the schema file and its comments.

There is also a manpage for racksdb CLI.