refactor(core): introduce notion of DB views
This commit introduces the notion of generic DB views, independant of
the presenter (CLI or whatever else). The generic package has a new
views module that defines 3 classes for handling DBViewSet, DBViews and
DBFilters definitions.

The RacksDB views modules define the views for RacksDB database. The
exec module now iterates over the views to define the CLI arguments with
a generic method to dump the requested content with optional filters.