refactor(web): split views in blueprint
The REST API views are now handled by a new class
RacksDBRESTAPIBlueprint that inherits from Flask Blueprint. The
RacksDBRESTAPIApp is the actual Flask application. It essentially
handles the CLI arguments and registers the blueprint.

This design allows other applications to register RacksDB blueprint for
full integration of its REST API.