fix(draw): display row front even when reversed
With data center aisle containment, one row might be reversed relative
to the other. This is correctly represented with
`RacksRow.reversed=true` when drawing the room, but not when drawing the
infrastructure. In the latter, the drawing will display the front of one
row and the rear of the other.

Fix this by displaying the racks in the correct order to always display
the row front in the infrastructure.