feat(draw): introduce drawing parameters
RacksDB generated room and infrastructure drawings can now be tuned with
some drawing parameters, replacing the respective hardcoded class
attributes that were used.

With the CLI, the drawing parameters are loaded from a YAML file
whose path is specified by --parameters optional. Alternatively,
parameters can be loaded from stdin. When the option is not used,
default drawing parameters defined in schema are used.

With the REST API, the drawing parameters must be provided in /draw
requests body either as YAML or JSON data. The HTTP verb for this route
has been changed from GET to POST to avoid unpredictable behaviours with
request body in GET requests.

In both cases, parameters are loaded as a GenericDB and checked against
the drawings schema. The parameters database does not have to contain
all drawing parameters, only parameters to override from defaults
defined in schema have to be specified.