feat(ui): introduce web UI (#32)
This commit introduces a brand new web UI to explore RacksDB content in
a web browser. This feature is a technology preview in beta version.

The frontend of this UI is developed in Typescript with Vue.js
framework and Tailwind CSS UI toolkit. The backend is the existing
RacksDB REST API. Documentation to install and use this web UI is
planned, along with many improvements in terms of design and usability.