feat(draw): reactive text labels with pango
For both infrastructures and room, add text labels in graphical
representations with fully featured Pango library through GObject
introspection, in replacement of Cairo show_text() so-called toy API.

Pango notably provides the possibility to give the expected size of the
text labels. This is used to decrease the font size until its fits into
the corresponding equipment or rack.