feat(draw): optional pixel perfact draws
This commit introduces the general pixel_perfect boolean drawing
parameters. When enabled, RacksDB properly align graphical
representation with pixels matrix. When enabled, concession is done on
relative dimensions correctness because:

- Lengths are rounded to integer to match exact pixels,
- Successive elements (eg. racks units) are always represented with the
  same size in pixels for visual consistency which could impact higher
  dimensions (eg. racks height that are calculated as multiple of these
  smaller height to match perfectly).

For example, the ratio between rack height and width (or equipments)
might not exactly match reality when this feature is enabled.

Equipments chassis are represented in the below equipments in racks to
fill potential gaps between the room inside racks and the multiple of
equipment width with the number of equipment in rack slot.

fix #4