feat(draw): ghost unselected equipment
This commit introduces the infrastructure ghost_unselected boolean
drawing parameter to control if unselected equipment is represented as
ghosted in racks to mark their presence in the infrastructure. When
disabled, the unselected equipment is simply discarded. This is disabled
by default.

Also add colors equipment ghost drawing parameter to specify the color
of ghosted equipment.