feat(web): add coordinates parameter in REST API
Add coordinates parameter on draw action for REST API that is
transposed as coordinates query parameter with allowed empty value on
draw endpoint. The coordinates_format query parameter was also added by
previous commit.

When this coordinates query parameter is used, REST API generates a
multipart response with both the diagram image file and the coordinates
file in the same response.

An optional dependency is added on popular requests-toolbelt on web
variant in order to generate the multipart response.

fix #56