fix(draw): make _equipment_width_slot real int
The result of a division in Python is a float. The type of
equipment.type.width is the rack_width defined type which asserts its
value is a part of a whole one, thus making its inverted value an whole
number which can be safely converted into an integer without rounding.