fix(core): dump optional but overriden properties
Properly dump properties in JSON and YAML when not defined in database
but overriden by specialized classes.

When a property is overriden by specialized class, it is renamed with
In this case, the property LOADED_PREFIX and dumpers handled this case
by calling the specialized method instead. However, when the property is
optional in database schema and not defined in database, the property
with LOADED_PREFIX is not defined. Thus, iterations over vars() actually
miss the call on specialized class in this case. This is fixed by
iterating over the properties of the object as defined in schema,
instead of relying on vars() function.