refactor(core): computed properties in schema
Instead of relying on a list of properties names defined in base class,
all computed properties are now expected to be declared in schema with
the computed boolean attribute set to True. Errors are raised if
computed properties are actually defined by user in database. This way,
the type of computed properties is clearly defined and it will clearly
appear in documentation.