feat(schema): equipment position computed property
Add position computed property on Node, StorageEquipment,
NetworkEquipment and MiscEquipment objects in schema to get their exact
position in the racks, in both width and height.

The new property is cached to avoid multiple instanciation of identical
RacksDBEquipmentPosition objects with successive calls of the property
on the same equipment object.

fix #72