feat(ui): infrastructure equipment filters
This commit adds the possibility to filter equipment in infrastructure
view. A filters slider component is introduced to set filters on the
rack name, equipment category, tags, equipment name. Another filters bar
component is introduced to represent the set of applied filters with the
possibilty to quickly remove them.

fix #42